+++ cmd/frontend/enterprise/BUILD.bazel
@@ -7,7 +7,7 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//cmd/frontend/graphqlbackend",
-        "//cmd/frontend/webhooks",
+        "//cmd/frontend/internal/webhooks",
         "//internal/api",
         "//internal/auth",
         "//internal/codeintel/types",
