+++ cmd/frontend/internal/authz/webhooks/BUILD.bazel
@@ -7,7 +7,7 @@ go_library(
     importpath = "github.com/sourcegraph/sourcegraph/cmd/frontend/internal/authz/webhooks",
     visibility = ["//cmd/frontend:__subpackages__"],
     deps = [
-        "//cmd/frontend/webhooks",
+        "//cmd/frontend/internal/webhooks",
         "//internal/api",
         "//internal/authz/permssync",
         "//internal/database",
@@ -29,7 +29,7 @@ go_test(
         "requires-network",
     ],
     deps = [
-        "//cmd/frontend/webhooks",
+        "//cmd/frontend/internal/webhooks",
         "//internal/api",
         "//internal/authz/permssync",
         "//internal/conf",
