+++ cmd/frontend/internal/batches/webhooks/bitbucketserver.go
@@ -10,7 +10,7 @@ import (
 	gh "github.com/google/go-github/v55/github"
 	sglog "github.com/sourcegraph/log"
 
-	fewebhooks "github.com/sourcegraph/sourcegraph/cmd/frontend/webhooks"
+	fewebhooks "github.com/sourcegraph/sourcegraph/cmd/frontend/internal/webhooks"
 	"github.com/sourcegraph/sourcegraph/internal/actor"
 	"github.com/sourcegraph/sourcegraph/internal/batches/store"
 	"github.com/sourcegraph/sourcegraph/internal/database"
