+++ cmd/frontend/internal/httpapi/webhookhandlers/handle_repo_authz_event.go
@@ -7,7 +7,7 @@ import (
 	gh "github.com/google/go-github/v55/github"
 	"github.com/sourcegraph/log"
 
-	"github.com/sourcegraph/sourcegraph/cmd/frontend/webhooks"
+	"github.com/sourcegraph/sourcegraph/cmd/frontend/internal/webhooks"
 	"github.com/sourcegraph/sourcegraph/internal/actor"
 	"github.com/sourcegraph/sourcegraph/internal/api"
 	"github.com/sourcegraph/sourcegraph/internal/authz"
