+++ cmd/frontend/internal/httpapi/webhookhandlers/handlers.go
@@ -3,7 +3,7 @@ package webhookhandlers
 import (
 	"github.com/sourcegraph/log"
 
-	"github.com/sourcegraph/sourcegraph/cmd/frontend/webhooks"
+	"github.com/sourcegraph/sourcegraph/cmd/frontend/internal/webhooks"
 	"github.com/sourcegraph/sourcegraph/internal/authz"
 )
