+++ cmd/frontend/internal/repos/webhooks/BUILD.bazel
@@ -10,7 +10,7 @@ go_library(
     deps = [
         "//cmd/frontend/enterprise",
         "//cmd/frontend/internal/repos/webhooks/resolvers",
-        "//cmd/frontend/webhooks",
+        "//cmd/frontend/internal/webhooks",
         "//internal/api",
         "//internal/codeintel",
         "//internal/conf/conftypes",
@@ -39,7 +39,7 @@ go_test(
         "requires-network",
     ],
     deps = [
-        "//cmd/frontend/webhooks",
+        "//cmd/frontend/internal/webhooks",
         "//internal/api",
         "//internal/conf",
         "//internal/database",
